rotator/function/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
azure-keyvault-keys
